backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.6.3
pydantic[email]==2.6.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
sentence-transformers==2.7.0
torch==2.2.2
requests==2.31.0
python-multipart==0.0.9
